+++ src/plotting/__init__.py
@@ -1,2 +1,3 @@
 from .plot_dfs import dpfsresult2plot
+from .plot_hfs import hfsresult2plot
 from .plot_upm import upmresult2plot
